Refactor directory structure to prepare for CSS Modules
