+++ src/components/Mission/index.jsx
@@ -2,10 +2,10 @@ import bindAll from 'lodash.bindall';
 import React from 'react';
 import Radium from 'radium';
 import update from 'react-addons-update';
-import Console from './Console';
-import Field from './Field';
-import MissionData from './MissionData';
-import MissionStyle from './MissionStyle';
+import Console from '../Console';
+import Field from '../Field';
+import Data from './Data';
+import styles from './style';
 
 class Mission extends React.Component {
   static get STATUS() {
@@ -47,7 +47,7 @@ class Mission extends React.Component {
 
   // TODO JSON APIサーバーを実装してそこから読み込むようにする
   handleParamsChange(params) {
-    const data = new MissionData(params).build().map((grid) =>
+    const data = new Data(params).build().map((grid) =>
       update(grid, { $merge: { isRevealed: false } })
     );
 
@@ -121,7 +121,7 @@ class Mission extends React.Component {
 
   render() {
     return (
-      <div style={MissionStyle.base}>
+      <div style={styles.base}>
         <Console
           onParamsChange={this.handleParamsChange}
           params={this.state.params}
