+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Mission from './Mission';
+import Mission from './components/Mission';
 
 ReactDOM.render(
   <Mission />,
